Migrate Checked property to not be portal-exclusive
