+++ GDEdit/GDEdit/GDEdit.csproj
@@ -180,7 +180,7 @@
     <Compile Include="Utilities\Objects\GeometryDash\LevelObjects\SpecialObjects\Portals\BlueMirrorPortal.cs" />
     <Compile Include="Utilities\Objects\GeometryDash\LevelObjects\SpecialObjects\Portals\YellowMirrorPortal.cs" />
     <Compile Include="Utilities\Objects\GeometryDash\LevelObjects\SpecialObjects\Portals\YellowGravityPortal.cs" />
-    <Compile Include="Utilities\Objects\GeometryDash\LevelObjects\SpecialObjects\Portals\IHasCheckedProperty.cs" />
+    <Compile Include="Utilities\Objects\GeometryDash\LevelObjects\SpecialObjects\IHasCheckedProperty.cs" />
     <Compile Include="Utilities\Objects\GeometryDash\LevelObjects\SpecialObjects\Portals\GamemodePortals\SpiderPortal.cs" />
     <Compile Include="Utilities\Objects\GeometryDash\LevelObjects\SpecialObjects\Portals\GamemodePortals\RobotPortal.cs" />
     <Compile Include="Utilities\Objects\GeometryDash\LevelObjects\SpecialObjects\Portals\GamemodePortals\WavePortal.cs" />
