+++ GDEdit/GDEdit/Utilities/Objects/GeometryDash/LevelObjects/SpecialObjects/OrbPad.cs
@@ -11,7 +11,7 @@ using System.Threading.Tasks;
 namespace GDEdit.Utilities.Objects.GeometryDash.LevelObjects.SpecialObjects
 {
     /// <summary>Represents an orb or a pad object.</summary>
-    public abstract class OrbPad : ConstantIDSpecialObject
+    public abstract class OrbPad : ConstantIDSpecialObject, IHasCheckedProperty
     {
         /// <summary>Represents the Switch Player Direction property of the orb or pad.</summary>
         [ObjectStringMappable(ObjectParameter.SwitchPlayerDirection)]
@@ -19,6 +19,13 @@ namespace GDEdit.Utilities.Objects.GeometryDash.LevelObjects.SpecialObjects
         {
             get => SpecialObjectBools[0];
             set => SpecialObjectBools[0] = value;
+        }
+        /// <summary>Represents the Checked property of the orb or pad.</summary>
+        [ObjectStringMappable(ObjectParameter.SpecialObjectChecked)]
+        public bool Checked
+        {
+            get => SpecialObjectBools[1];
+            set => SpecialObjectBools[1] = value;
         }
 
         /// <summary>Initializes a new instance of the <seealso cref="OrbPad"/> class.</summary>
